Get rid of wrapper DOM node
And fix some issues with setting classes on the editor node

BREAKING: The editor view no longer wraps its editable DOM element in
a wrapper element. The `ProseMirror` CSS class now applies directly to
the editable element. The `ProseMirror-content` CSS class is still
present for ease of upgrading but will be dropped in the next release.