Give up on setting a DOM selection for node selections
Browsers will do stupid stuff when trying to set a selection
between non-editable nodes, the wrapper nodes were problematic
and didn't really work, and Firefox doesn't make it possible to
suppress selection styling on images inside a selection.

Now the editor just sets the selection to empty while keeping
the focus on the editor when there is a node selection.