Replace handleDOMEvent with handleDOMEvents prop
Automatically register the necessary handlers.

BREAKING: The `handleDOMEvent` prop has been dropped in favor of the
[`handleDOMEvents`](##view.EditorProps.handleDOMEvents) (plural) prop.

FEATURE: Added support for a [`handleDOMEvents`
prop](##view.EditorProps.handleDOMEvents), which allows you to provide
handler functions per DOM event, and works even for events that the
editor doesn't normally add a handler for.