Move library from actions to transactions
Tests not updated yet

BREAKING: The `onChange` prop has been replaced by a
[`dispatchTransaction`](##view.EditorProps.dispatchTransaction) prop
(which takes a transaction instead of an action).

FEATURE: Add view method [`dispatch`](##view.EditorView.dispatch),
which provides a convenient way to dispatch transactions.

FEATURE: The
[`dispatchTransaction`](##view.EditorProps.dispatchTransaction) (used
to be `onAction`) prop is now optional, and will default to simply
applying the transaction to the current view state.