Bring back DOM selections for node selections
Browsers will scroll to the top of the editor when the editable node
is focused but no selection is active, and will reset the selection to
the start in various situations (such as re-focusing the window).

Work around Webkit's limitations on the start/end of the selection by
briefly setting the element editable.