Add a side option for widget decorations
BREAKING: The `associative` option to widget decorations is no longer
supported. To make a widget left-associative, set its `side` option to
a negative number. `associative` will continue to work with a warning
until the next release.

FEATURE: [Widget decorations](##view.Decoration^widget) now support a
`side` option that controls which side of them the cursor is drawn,
where they move when content is inserted at their position, and the
order in which they appear relative to other widgets at the same
position.