Allow the DOMChange timeout to be customized
This 20ms timeout is to fix issues with Android Chrome, but it makes a
jitter in the editor when plugins cancel/modify user input.

This change allows the implementor to change the timeout. This may be
done statically if Android is not a supported platform, or the
implementor may use means to detect the platform and change the timeout
based on that.