Fix explosion on creating editor when editable returns false.
The SelectionReader is created in the EditorView constructor before calling 'initInput'.

If the view isn't editable, this starts a poller to keep the selection in sync with the DOM, which ends up calling this.view.domObserver.flush().

Since the domObserver gets created on the view on line 49 of prosemirror-view index.js, this doesn't yet exist, so boom.

I'm not quite sure why my usage is triggering this bug, I assume other people's isn't? I'm rendering this editor output on a tooltip which is getting moved about, so perhaps it's something to do with that - regardless, this seems like a clear bug to me. Creating the domObserver before the SelectionReader appears to fix it.

Note that the SelectionReader can also generate a transaction before init returns, which I'm not sure if is considered a problem in ProseMirror.

Alternatively it may be better to disable the selection api while the editor isn't editable?

FIX: Initializing the editor view non-editable no longer causes a crash.