Improve tracking of shift status
Use `shiftKey` property from keydown and mousedown events to avoid
getting stuck with the wrong value when the editor is entered/left
with shift active.

FIX: Fix an issue where the editor got stuck believing shift was down
(and hence pasting as plain text) when it was unfocused with shift
held down.

Issue prosemirror/prosemirror#834