Work around IE11 corrupting the DOM selection after backspacing
When you press backspace after the first character in a textblock,
it'd force-move the selection after applying that change, sometimes to
the end of the textblock, sometimes to the end of the document.

This ignores such selection changes and forces the selection back to
where it should be.

FIX: Fix an issue where backspacing out the first character in a
textblock would cause IE11 to move the selection to some incorrect
position.

See https://discuss.prosemirror.net/t/backspace-or-canc-on-ie11/2053