Use fallback in setSelection to fix chrome bug
In some cases with Chrome the selection is empty after calling
collapse, even when it should be valid. This appears to be a bug,
but it is difficult to isolate. If this happens fallback to the old
path without using extend.

FIX: Fix a crash on Chrome during selection updates when `Selection.collapse` inexplicably leaves the selection empty.